virtcontainers: Define a Network interface
And move the Linux implementation into a GOOS specific file.

Fixes #3005
