+++ src/runtime/virtcontainers/acrn.go
@@ -356,7 +356,7 @@ func (a *Acrn) setConfig(config *HypervisorConfig) error {
 }
 
 // CreateVM is the VM creation
-func (a *Acrn) CreateVM(ctx context.Context, id string, network *Network, hypervisorConfig *HypervisorConfig) error {
+func (a *Acrn) CreateVM(ctx context.Context, id string, network Network, hypervisorConfig *HypervisorConfig) error {
 	// Save the tracing context
 	a.ctx = ctx
